refactor: streamline imports and update logo assets; replace logo.png with logo.svg
