+++ src/assets/index.js
@@ -1,0 +1,14 @@
+export { default as Ai } from "./ai.png";
+export { default as Atlassian } from "./atlassian.png";
+export { default as Blog01 } from "./blog01.png";
+export { default as Blog02 } from "./blog02.png";
+export { default as Blog03 } from "./blog03.png";
+export { default as Blog04 } from "./blog04.png";
+export { default as Blog05 } from "./blog05.png";
+export { default as Google } from "./google.png";
+export { default as Dropbox } from "./dropbox.png";
+export { default as Logo } from "./logo.svg";
+export { default as Shopify } from "./shopify.png";
+export { default as Slack } from "./slack.png";
+export { default as People } from "./people.png";
+export { default as PossibilityImg } from "./possibility.png";
