+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as Article } from "./article/Article.jsx";
+export { default as Brand } from "./brand/Brand.jsx";
+export { default as CTA } from "./CTA/CTA.jsx";
+export { default as Feature } from "./feature/Feature.jsx";
+export { default as Navbar } from "./navbar/Navbar.jsx";
