+++ src/containers/blog/Blog.jsx
@@ -1,9 +1,5 @@
-import Article from "../../components/article/Article"
-import blog1 from "../../assets/blog01.png"
-import blog2 from "../../assets/blog02.png"
-import blog3 from "../../assets/blog03.png"
-import blog4 from "../../assets/blog04.png"
-import blog5 from "../../assets/blog05.png"
+import { Article } from "../../components"
+import { Blog01, Blog02, Blog03, Blog04, Blog05 } from "../../assets"
 
 import "./blog.css"
 
@@ -17,29 +13,29 @@ const Blog = () => {
             <div className="gpt__blog-container">
                 <div className="gpt__blog-container__groupA">
                     <Article
-                        imgUrl={blog1}
+                        imgUrl={Blog01}
                         date="Sep 26, 2021"
                         text="GPT-3 and Open AI is the future. Let us explore how it is?"
                     />
                 </div>
                 <div className="gpt__blog-container__groupB">
                     <Article
-                        imgUrl={blog2}
+                        imgUrl={Blog02}
                         date="Sep 26, 2021"
                         text="GPT-3 and Open AI is the future. Let us explore how it is?"
                     />
                     <Article
-                        imgUrl={blog3}
+                        imgUrl={Blog03}
                         date="Sep 26, 2021"
                         text="GPT-3 and Open AI is the future. Let us explore how it is?"
                     />
                     <Article
-                        imgUrl={blog4}
+                        imgUrl={Blog04}
                         date="Sep 26, 2021"
                         text="GPT-3 and Open AI is the future. Let us explore how it is?"
                     />
                     <Article
-                        imgUrl={blog5}
+                        imgUrl={Blog05}
                         date="Sep 26, 2021"
                         text="GPT-3 and Open AI is the future. Let us explore how it is?"
                     />
