+++ src/containers/index.js
@@ -1,0 +1,6 @@
+export { default as Blog } from "./blog/Blog.jsx";
+export { default as Features } from "./features/Features.jsx";
+export { default as Footer } from "./footer/Footer.jsx";
+export { default as Header } from "./header/Header.jsx";
+export { default as Possibility } from "./possibility/Possibility.jsx";
+export { default as WhatGPT } from "./what-gpt/WhatGPT.jsx";
